Apply intermediate step in removing Voyager Navigation
Screens are still classes, instead of composable functions.

Need to figure out how to specify the task to display (when clicked from HomeScreen) with event/call-back, instead of passing the task to display directly.

But: As intermediate step, get the TaskScreen to display the correct task when clicked by using param.